+++ src/lifish/entities/BreakableWall.cpp
@@ -25,12 +25,12 @@ using lif::BreakableWall;
 using lif::TILE_SIZE;
 
 BreakableWall::BreakableWall(const sf::Vector2f& pos, const std::string& texture_name,
-		unsigned short life, unsigned int score)
+		unsigned life, unsigned int score)
 	: lif::Entity(pos)
 {
 	_setupComponents(life, score);
 	auto& animation = _setupAnimations(texture_name);
-	for (unsigned short i = 0; i < 4; ++i)
+	for (unsigned i = 0; i < 4; ++i)
 		animation.addFrame(sf::IntRect(
 					i * TILE_SIZE, 
 					0,
@@ -41,12 +41,12 @@ BreakableWall::BreakableWall(const sf::Vector2f& pos, const std::string& texture
 }
 
 BreakableWall::BreakableWall(const sf::Vector2f& pos, 
-		const unsigned short id, unsigned short life)
+		const unsigned short id, unsigned life)
 	: lif::Entity(pos)
 {
 	_setupComponents(life, lif::conf::wall::breakable::VALUE);
 	auto& animation = _setupAnimations(lif::getAsset("graphics", "breakable.png"));	
-	for (unsigned short i = 0; i < 4; ++i)
+	for (unsigned i = 0; i < 4; ++i)
 		animation.addFrame(sf::IntRect(
 					i * TILE_SIZE, 
 					(id-1) * TILE_SIZE, 
@@ -55,7 +55,7 @@ BreakableWall::BreakableWall(const sf::Vector2f& pos,
 	animated->setAnimation(animation);
 }
 
-void BreakableWall::_setupComponents(unsigned short life, unsigned int score) {
+void BreakableWall::_setupComponents(unsigned life, unsigned int score) {
 	addComponent(new lif::Fixed(*this));
 	addComponent(new lif::Scored(*this, score));
 	addComponent(new lif::Lifed(*this, life));
