+++ src/lifish/entities/BreakableWall.hpp
@@ -19,13 +19,13 @@ class BreakableWall : public lif::Entity {
 protected:
 	/** Constructor used by TransparentWall */
 	BreakableWall(const sf::Vector2f& pos, const std::string& texture_name, 
-			unsigned short life, unsigned int score);
+			unsigned life, unsigned int score);
 
 	Animation& _setupAnimations(const std::string& texture_name);
-	void _setupComponents(unsigned short life, unsigned int score);
+	void _setupComponents(unsigned life, unsigned int score);
 
 public:
-	explicit BreakableWall(const sf::Vector2f& pos, const unsigned short id, unsigned short life = 1);
+	explicit BreakableWall(const sf::Vector2f& pos, const unsigned short id, unsigned life = 1);
 };
 
 }
