+++ src/lifish/entities/FreeBullet.cpp
@@ -33,12 +33,12 @@ FreeBullet::FreeBullet(const sf::Vector2f& pos, lif::Angle angle,
 	addComponent(new lif::Drawable(*this, *animated));
 	auto& a_move = animated->addAnimation("move");
 	auto& a_destroy = animated->addAnimation("destroy");
-	for (unsigned short i = 0; i < data.nMotionFrames; ++i) 
+	for (unsigned i = 0; i < data.nMotionFrames; ++i) 
 		a_move.addFrame(sf::IntRect(
 				i * TILE_SIZE,
 				(info.id - 101) * TILE_SIZE,
 				TILE_SIZE, TILE_SIZE));
-	for (unsigned short i = 0; i < data.nDestroyFrames; ++i)
+	for (unsigned i = 0; i < data.nDestroyFrames; ++i)
 		a_destroy.addFrame(sf::IntRect(
 				(data.nMotionFrames + i) * TILE_SIZE,
 				(info.id - 101) * TILE_SIZE,
